Memisahkan root dan vuex
